Chore: Comprehensive linting fixes across the project
This commit addresses a large number of ESLint errors and warnings identified in your project, focusing on improving code quality, maintainability, and adherence to React best practices.

Key changes include:

1.  **Critical Errors Resolved:**
    *   I fixed a parsing error in `src/components/EarthquakeSequenceChart.test.jsx`.
    *   I resolved 9 `react-hooks/rules-of-hooks` errors in `src/components/EarthquakeSequenceChart.jsx` by ensuring hooks are called unconditionally.

2.  **Unused Variables (`no-unused-vars`):**
    *   I removed or addressed approximately 11 instances of unused variables across various components and test files. I encountered some challenges with `EarthquakeSequenceChart.test.jsx`, but most issues were resolved.

3.  **Accessibility (`jsx-a11y`):**
    *   I improved accessibility in `ClusterDetailModal.jsx` and `EarthquakeDetailView.jsx` by:
        *   Refactoring click handlers and keyboard interactions for modal backdrops and content.
        *   Using targeted ESLint disable comments where necessary to resolve conflicts between `jsx-a11y` rules and `axe-core` `nested-interactive` violations for modal dialog patterns.

4.  **React Hook Dependencies (`react-hooks/exhaustive-deps`):**
    *   I addressed all 15 warnings by:
        *   Adding missing dependencies to `useEffect` and `useCallback` hooks.
        *   Memoizing values with `useMemo` and `useCallback` to prevent unnecessary re-renders and effect executions.
        *   Removing unnecessary dependencies.
        *   I significantly refactored a `useEffect` in `EarthquakeDataContext.jsx` by splitting it into two hooks (one for initial load, one for refresh interval) and using `useCallback`. This resolved the lint warning and a related test failure.

5.  **React Fast Refresh (`react-refresh/only-export-components`):**
    *   I improved compliance by moving non-component exports (constants, helpers, reducers, context objects) from component/context files (`SimplifiedDepthProfile.jsx`, `EarthquakeDataContext.jsx`, `UIStateContext.jsx`) to separate utility files (e.g., `earthquakeDataContextUtils.js`).
    *   A persistent ESLint warning for this rule on the Provider component exports in the context files is noted but accepted for now, as the primary structure has been improved and tests pass.

6.  **Minor Cleanup:**
    *   An unused `eslint-disable` directive was removed.

Most pre-existing test failures were not addressed as part of this linting-focused effort, but test failures introduced by these changes (e.g., in `EarthquakeDataContext.jsx`) were resolved.
