+++ package.json
@@ -30,7 +30,7 @@
   },
   "devDependencies": {
     "@axe-core/react": "^4.10.1",
-    "@eslint/js": "^9.28.0",
+    "@eslint/js": "^9.29.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
