+++ src/components/ClusterDetailModal.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import ClusterMiniMap from './ClusterMiniMap'; // Added import for the mini-map
 import EarthquakeSequenceChart from './EarthquakeSequenceChart'; // Import the new chart
-import { getMagnitudeColor } from '../utils/utils.js'; // Corrected import for getMagnitudeColor
+// import { getMagnitudeColor } from '../utils/utils.js'; // Corrected import for getMagnitudeColor - Unused
 
 /**
  * A modal component to display detailed information about an earthquake cluster.
@@ -114,18 +114,22 @@ function ClusterDetailModal({ cluster, onClose, formatDate, getMagnitudeColorSty
     const depthRangeStr = depthDataAvailable ? `${minDepth.toFixed(1)}km - ${maxDepth.toFixed(1)}km` : 'N/A';
 
     return (
+        // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
         <div
             className="fixed inset-0 bg-slate-900 bg-opacity-75 flex items-center justify-center z-[51] p-4 transition-opacity duration-300 ease-in-out"
-            onClick={onClose} // Removed backdrop click to close, rely on Esc key and close button
-            role="dialog"
-            aria-modal="true"
-            aria-labelledby="cluster-detail-title"
+            onClick={(e) => { if (e.target === e.currentTarget) { onClose(); } }}
+            // Removed role="button", tabIndex, onKeyDown, and aria-label from backdrop to avoid nested-interactive
+            // Escape key is handled by global event listener and focus trap.
         >
             <div
                 ref={modalContentRef}
-                tabIndex="-1" // Make modal container focusable for trap if no inner elements are
+                tabIndex="-1"
+                role="dialog" // This is the actual dialog panel
+                aria-modal="true" // Indicates this is a modal dialog
+                aria-labelledby="cluster-detail-title" // Associates with the title
+                // onKeyDown={(e) => { if (e.key === 'Escape') { onClose(); e.stopPropagation(); } }} // Removed as Escape is handled by useEffect
                 className="bg-slate-800 p-4 sm:p-6 rounded-lg shadow-2xl w-full max-w-2xl max-h-[90svh] flex flex-col border border-slate-700 overflow-y-auto scrollbar-thin scrollbar-thumb-slate-600 scrollbar-track-slate-700"
-                onClick={e => e.stopPropagation()} // Prevent backdrop click from triggering inside modal
+                // onClick={e => e.stopPropagation()} // This was already removed
             >
                 {/* Header */}
                 <div className="flex items-center justify-between pb-3 border-b border-slate-700 mb-4">
@@ -183,7 +187,7 @@ function ClusterDetailModal({ cluster, onClose, formatDate, getMagnitudeColorSty
                             const originalClassName = `w-full text-left p-2.5 rounded-md border ${getMagnitudeColorStyle ? getMagnitudeColorStyle(quake.properties?.mag) : 'bg-slate-700 border-slate-600'} hover:border-slate-500 transition-colors cursor-pointer focus:outline-none focus:ring-2 focus:ring-indigo-400`;
 
                             return (
-                            // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-noninteractive-element-interactions, jsx-a11y/no-static-element-interactions
+                            // The eslint-disable-next-line previously here was removed as it was unused.
                             <button
                                 key={quake.id}
                                 type="button" // Explicit type for button
