+++ src/components/SimplifiedDepthProfile.jsx
@@ -1,30 +1,7 @@
 // src/SimplifiedDepthProfile.jsx
 import React from 'react';
 import { getMagnitudeColor } from '../utils/utils.js';
-
-export const DEPTH_COMPARISONS = [ // Added export
-  { name: "Burj Khalifa", depth: 0.828 },
-  { name: "Krubera Cave (deepest cave)", depth: 2.197 },
-  { name: "Grand Canyon (average depth)", depth: 1.83 },
-  { name: "Challenger Deep (ocean deepest)", depth: 10.935 },
-  { name: "Average Continental Crust", depth: 35 },
-  { name: "Height of Mount Everest", depth: 8.848, isHeight: true },
-  { name: "Typical Commercial Flight Altitude", depth: 10.6, isHeight: true },
-  { name: "Depth of Titanic Wreckage", depth: 3.8 },
-  { name: "Deepest Gold Mine (Mponeng, South Africa)", depth: 4.0 },
-  { name: "Average Ocean Depth", depth: 3.7 },
-  { name: "Kola Superdeep Borehole (deepest artificial point)", depth: 12.262 },
-  { name: "Deepest Point in the Arctic Ocean (Molloy Deep)", depth: 5.55 },
-  { name: "Deepest Point in the Atlantic Ocean (Puerto Rico Trench)", depth: 8.376 },
-  { name: "Deepest Point in the Indian Ocean (Java Trench)", depth: 7.725 },
-  { name: "Typical Geothermal Well Depth", depth: 2.0 },
-  { name: "Depth of Lake Baikal (deepest lake)", depth: 1.642 },
-  { name: "Panama Canal Max Depth", depth: 0.018 },
-  { name: "Suez Canal Max Depth", depth: 0.024 },
-  { name: "Shallow Focus Earthquakes (Upper Limit)", depth: 70 },
-  { name: "Intermediate Focus Earthquakes (Upper Limit)", depth: 300 },
-  { name: "Deep Focus Earthquakes (Upper Limit)", depth: 700 },
-];
+import { DEPTH_COMPARISONS, getDynamicContextualComparisons } from './simplifiedDepthProfileUtils.js';
 
 /**
  * Displays a simplified, illustrative vertical profile of Earth's layers
@@ -296,110 +273,5 @@ function SimplifiedDepthProfile({ earthquakeDepth, magnitude }) {
   // This call is already moved up before 'layers' was used by it, so this comment is slightly misplaced in search.
   // The actual call to getDynamicContextualComparisons is correctly placed higher now.
 
-
-// Helper function implementation (outside the component for clarity)
-export function getDynamicContextualComparisons(currentDepth, comparisonsList) { // Added export
-  // EarthLayers parameter removed as per new requirements
-
-  const RELATABLE_OBJECTS = [
-    { name: "Eiffel Towers", height: 0.3, singular: "Eiffel Tower" },
-    // Burj Khalifa is in comparisonsList, so we can use its data directly.
-    // Mount Everest is in comparisonsList (isHeight: true), use its data.
-  ];
-
-  const depth = parseFloat(currentDepth);
-  if (isNaN(depth)) { // Should have been caught before, but good to be safe.
-    return ["Depth information is currently unavailable."];
-  }
-
-  const userFriendlyBenchmarks = comparisonsList
-    .filter(c => !c.isHeight && !c.name.includes("Focus Earthquakes"))
-    .sort((a, b) => a.depth - b.depth);
-
-  // 1. Check for "very close" match to any user-friendly benchmark
-  for (const benchmark of userFriendlyBenchmarks) {
-    const difference = Math.abs(depth - benchmark.depth);
-    let isClose = false;
-    if (benchmark.depth > 1) { // Use percentage for larger benchmarks
-      isClose = difference <= benchmark.depth * 0.10; // 10% threshold for "nearly as deep"
-    } else { // Use absolute for smaller benchmarks (e.g. canals, Burj)
-      isClose = difference <= 0.1; // 100m absolute difference
-    }
-
-    if (isClose) {
-      return [`${depth.toFixed(1)} km is nearly as deep as the ${benchmark.name} (${benchmark.depth.toFixed(1)} km)!`];
-    }
-  }
-
-  // 2. Check if "even further down than" a significant, well-known benchmark
-  // Prioritize deeper, more impressive benchmarks for this message.
-  // Select a few well-known items that are good for this "further down than" comparison.
-  const significantBenchmarks = userFriendlyBenchmarks.filter(b =>
-    b.name === "Kola Superdeep Borehole (deepest artificial point)" ||
-    b.name === "Challenger Deep (ocean deepest)" ||
-    b.name === "Average Continental Crust" || // Good mid-range geological feature
-    b.name === "Deepest Gold Mine (Mponeng, South Africa)"
-  ).sort((a, b) => b.depth - a.depth); // Sort deepest first
-
-  for (const benchmark of significantBenchmarks) {
-    if (depth > benchmark.depth && depth < benchmark.depth * 3) { // If deeper, but not astronomically deeper
-      return [`That's incredibly deep! It's even further down than the ${benchmark.name} (${benchmark.depth.toFixed(1)} km)!`];
-    }
-  }
-
-  // 3. Use a relatable analogy for very significant depths
-  let analogyMessage = "";
-
-  const mountEverest = comparisonsList.find(c => c.name === "Height of Mount Everest"); // isHeight = true
-  const burjKhalifa = comparisonsList.find(c => c.name === "Burj Khalifa");
-
-  if (depth > 5 && mountEverest) { // Use Mount Everest for reasonably large depths
-      const numObjects = Math.round(depth / mountEverest.depth);
-      if (numObjects > 1) {
-        analogyMessage = `Wow, ${depth.toFixed(1)} km is a long way down – that's like stacking about ${numObjects} Mount Everests on top of each other!`;
-      } else if (numObjects === 1) {
-        analogyMessage = `Wow, ${depth.toFixed(1)} km is a long way down – that's about as deep as Mount Everest is tall!`;
-      }
-  } else if (depth > 1 && burjKhalifa) { // Use Burj Khalifa for moderate depths
-      const numObjects = Math.round(depth / burjKhalifa.depth);
-       if (numObjects > 1) {
-        analogyMessage = `That's quite deep! ${depth.toFixed(1)} km is like stacking about ${numObjects} Burj Khalifas!`;
-      } else if (numObjects === 1 && depth > burjKhalifa.depth * 1.1) { // Only if significantly more than 1
-         analogyMessage = `That's quite deep! ${depth.toFixed(1)} km is like stacking about ${numObjects} Burj Khalifas!`;
-      } else if (numObjects === 1 ) { // if depth is between 0.828 and ~0.9km, it's roughly 1 Burj Khalifa
-         analogyMessage = `That's quite deep! ${depth.toFixed(1)} km is about the height of the Burj Khalifa!`;
-      }
-  } else if (depth > 0.5) { // Use Eiffel Tower for smaller depths
-      const eiffelTower = RELATABLE_OBJECTS.find(r => r.name === "Eiffel Towers");
-      const numObjects = Math.round(depth / eiffelTower.height);
-      if (numObjects >=1) {
-          analogyMessage = `That's pretty deep! ${depth.toFixed(1)} km is like stacking ${numObjects} ${numObjects === 1 ? eiffelTower.singular : eiffelTower.name}!`;
-      }
-  }
-
-  if (analogyMessage) {
-    return [analogyMessage];
-  }
-
-  // 4. Fallback messages
-  if (depth === 0) {
-    return ["This earthquake was right at the surface!"];
-  }
-  if (depth < 0.1 && depth > 0) {
-    return [`${depth.toFixed(2)} km is very close to the surface!`];
-  }
-  if (depth > 700) { // Still acknowledge exceptionally deep ones simply
-      return [`Whoa, ${depth.toFixed(0)} km is incredibly deep, way down into the Earth's mantle!`];
-  }
-  if (depth > 300) {
-      return [`That's a very deep earthquake, ${depth.toFixed(0)} km down!`];
-  }
-   if (depth > 70) {
-      return [`That's a deep earthquake, ${depth.toFixed(0)} km down!`];
-  }
-
-  // Generic fallback if no other condition met (e.g., shallow but not super shallow, no close comparisons)
-  return [`That's an earthquake at ${depth.toFixed(1)} km deep.`];
-}
 
 export default SimplifiedDepthProfile;
